Ajuste listar e cadastrar Avaliacao e Ajuda

+++ src/controller/ctlAjuda.js
@@ -1,33 +1,33 @@
 import * as db from "./firebase";
 const Swal = require('sweetalert2')
 
-var listarTodos = function listarTodos() {
-    db.ajuda.get().then((snapshot) => {
+var listarTodos = async function listarTodos() {
+    return await db.ajuda.get().then((snapshot) => {
         const data = snapshot.docs.map((doc) => ({
             id: doc.id,
             ...doc.data(),
         }));
-        console.log("Todas Avaliações = ", data);
+        return data;
     });
 }
 
-var bucarPorId = function bucarPorId(id) {
-    db.ajuda
+var bucarPorId = async function bucarPorId(id) {
+    return await db.ajuda
         .doc(id).get()
         .then((doc) => {
             if (!doc.exists) return;
-            console.log("Dados da Avaliação Selecionada = ", doc.data());
+            return doc.data();
         })
         .catch((error) => {
             console.error("Erro ao buscar Avaliação", error);
-            Swal.fire("Erro!", "A código da Ajuda fornecido não existe!", "error");
+            Swal.fire("Erro!", "O código da Ajuda fornecido não existe!", "error");
         });
 }
 
 var incluir = function incluir(dados) {
-    db.ajuda
+    return db.ajuda
         .add(dados)
-        .then((ref) => {
+        .then(() => {
             Swal.fire("Salvo!", "O pedido de ajuda foi salvo com sucesso!", "success");
         })
         .catch((error) => {
@@ -37,7 +37,7 @@ var incluir = function incluir(dados) {
 }
 
 var alterar = function alterar(id, dados) {
-    db.ajuda
+    return db.ajuda
         .doc(id)
         .update(dados)
         .then(() => {
@@ -50,16 +50,29 @@ var alterar = function alterar(id, dados) {
 }
 
 var excluir = function excluir(id) {
-    db.ajuda
-        .doc(id)
-        .delete()
-        .then(() => {
-            Swal.fire("Deletado!", "O pedido de ajuda foi deletado com sucesso!", "success");
-        })
-        .catch((error) => {
-            console.error("Erro ao excluir avaliação", error);
-            Swal.fire("Erro!", "Houve um problema ao tentar excluir o pedido de ajuda!", "error");
-        });
+    return Swal.fire({
+        title: "Atenção",
+        text: "Deseja realmente excluir o registro?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Sim, excluir!",
+        cancelButtonText: "Cancelar",
+    }).then((result) => {
+        if (result.isConfirmed) {
+            db.ajuda
+                .doc(id)
+                .delete()
+                .then(() => {
+                    Swal.fire("Deletado!", "O pedido de ajuda foi deletado com sucesso!", "success");
+                })
+                .catch((error) => {
+                    console.error("Erro ao excluir avaliação", error);
+                    Swal.fire("Erro!", "Houve um problema ao tentar excluir o pedido de ajuda!", "error");
+                });
+        }
+    });
 }
 
 export {
